+++ src/http.rs
@@ -1,4 +1,4 @@
-//! convertions to and from internal gateway types and http crate types
+//! conversions to and from internal gateway types and http crate types
 
 // Std
 use std::collections::HashMap;
@@ -7,7 +7,12 @@ use std::collections::HashMap;
 use body::Body;
 use request::{GatewayRequest, RequestContext};
 use response::GatewayResponse;
+use rust_http::header::CONTENT_TYPE;
 use rust_http::{Request as HttpRequest, Response as HttpResponse};
+use serde::de::value::Error as SerdeError;
+use serde::Deserialize;
+use serde_json;
+use serde_urlencoded;
 
 /// API gateway pre-parsed http query string parameters
 struct QueryStringParameters(HashMap<String, String>);
@@ -19,8 +24,57 @@ struct PathParameters(HashMap<String, String>);
 /// [stage variables](https://docs.aws.amazon.com/apigateway/latest/developerguide/stage-variables.html)
 struct StageVariables(HashMap<String, String>);
 
+/// Payload deserialization errors
+#[derive(Debug, Fail)]
+pub enum PayloadError {
+    /// Returned when `application/json` bodies fail to deserialize a payload
+    #[fail(display = "failed to parse payload from application/json")]
+    Json(serde_json::Error),
+    /// Returned when `application/x-www-form-urlencoded` bodies fail to deserialize a payload
+    #[fail(display = "failed to parse payload application/x-www-form-urlencoded")]
+    WwwFormUrlEncoded(SerdeError),
+}
+
 /// Extentions for `lando::Request` objects that
-/// provide access to API gateway features
+/// provide access to [API gateway features](https://docs.aws.amazon.com/apigateway/latest/developerguide/set-up-lambda-proxy-integrations.html#api-gateway-simple-proxy-for-lambda-input-format)
+///
+/// In addition, you can also access a request's body in deserialized format
+/// for payloads sent in `application/x-www-form-urlencoded` or
+/// `application/x-www-form-urlencoded` format
+///
+/// ```rust
+/// #[macro_use] extern crate cpython;
+/// #[macro_use] extern crate lando;
+/// #[macro_use] extern crate serde_derive;
+///
+/// use lando::{Response, RequestExt};
+///
+/// #[derive(Debug,Deserialize,Default)]
+/// struct Args {
+///   #[serde(default)]
+///   x: usize,
+///   #[serde(default)]
+///   y: usize
+/// }
+///
+/// # fn main() {
+/// gateway!(|request, _| {
+///   let args: Args = request.payload()
+///     .unwrap_or_else(|_parse_err| None)
+///     .unwrap_or_default();
+///   Ok(
+///      Response::new(
+///        format!(
+///          "{} + {} = {}",
+///          args.x,
+///          args.y,
+///          args.x + args.y
+///        )
+///      )
+///   )
+/// });
+/// # }
+/// ```
 pub trait RequestExt {
     /// Return pre-parsed http query string parameters
     /// associated with the API gateway request
@@ -33,9 +87,22 @@ pub trait RequestExt {
     fn stage_variables(&self) -> HashMap<String, String>;
     /// Return request context assocaited with the API gateway request
     fn request_context(&self) -> RequestContext;
+
+    /// Return the Result of a payload parsed into a serde Deserializeable
+    /// type
+    ///
+    /// Currently only `application/x-www-form-urlencoded`
+    /// and `application/json` flavors of content type
+    /// are supported
+    ///
+    /// A [PayloadError](enum.PayloadError.html) will be for undeserializable
+    /// payloads. If no body is provided `Ok(None)` will be returned.
+    fn payload<D>(&self) -> Result<Option<D>, PayloadError>
+    where
+        for<'de> D: Deserialize<'de>;
 }
 
-impl<T> RequestExt for HttpRequest<T> {
+impl RequestExt for HttpRequest<super::Body> {
     fn query_string_parameters(&self) -> HashMap<String, String> {
         self.extensions()
             .get::<QueryStringParameters>()
@@ -60,6 +127,26 @@ impl<T> RequestExt for HttpRequest<T> {
             .get::<RequestContext>()
             .map(|ext| ext.clone())
             .unwrap_or(Default::default())
+    }
+
+    fn payload<D>(&self) -> Result<Option<D>, PayloadError>
+    where
+        for<'de> D: Deserialize<'de>,
+    {
+        self.headers()
+            .get(CONTENT_TYPE)
+            .map(|ct| match ct.to_str() {
+                Ok("application/x-www-form-urlencoded") => {
+                    serde_urlencoded::from_bytes::<D>(self.body().as_ref())
+                        .map_err(PayloadError::WwwFormUrlEncoded)
+                        .map(Some)
+                }
+                Ok("application/json") => serde_json::from_slice::<D>(self.body().as_ref())
+                    .map_err(PayloadError::Json)
+                    .map(Some),
+                _ => Ok(None),
+            })
+            .unwrap_or_else(|| Ok(None))
     }
 }
 
@@ -107,12 +194,15 @@ impl From<GatewayRequest> for HttpRequest<Body> {
             request_context,
         } = value;
 
-        // build an http::Result from a lando::Request
+        // build an http::Request from a lando::Request
         let mut builder = HttpRequest::builder();
         builder.method(http_method.as_str()).uri({
             format!(
                 "https://{}{}",
-                headers.get("Host").unwrap_or(&String::new()),
+                headers
+                    .get("Host")
+                    .or_else(|| headers.get("host"))
+                    .unwrap_or(&String::new()),
                 path
             )
         });
@@ -142,14 +232,15 @@ impl From<GatewayRequest> for HttpRequest<Body> {
 #[cfg(test)]
 mod tests {
     use super::GatewayRequest;
-    use RequestExt;
+    use rust_http::header::{CONTENT_TYPE, HOST};
     use rust_http::Request as HttpRequest;
     use std::collections::HashMap;
+    use RequestExt;
 
     #[test]
     fn requests_convert() {
         let mut headers = HashMap::new();
-        headers.insert("Host".to_owned(), "www.rust-lang.org".to_owned());
+        headers.insert(HOST.as_str().to_string(), "www.rust-lang.org".to_owned());
         let gwr: GatewayRequest = GatewayRequest {
             path: "/foo".into(),
             http_method: "GET".into(),
@@ -167,7 +258,7 @@ mod tests {
     #[test]
     fn requests_have_query_string_ext() {
         let mut headers = HashMap::new();
-        headers.insert("Host".to_owned(), "www.rust-lang.org".to_owned());
+        headers.insert(HOST.as_str().to_string(), "www.rust-lang.org".to_owned());
         let mut query = HashMap::new();
         query.insert("foo".to_owned(), "bar".to_owned());
         let gwr: GatewayRequest = GatewayRequest {
@@ -179,5 +270,90 @@ mod tests {
         };
         let actual = HttpRequest::from(gwr);
         assert_eq!(actual.query_string_parameters(), query.clone());
+    }
+
+    #[test]
+    fn requests_have_form_post_parseable_payloads() {
+        let mut headers = HashMap::new();
+        headers.insert(HOST.as_str().to_string(), "www.rust-lang.org".to_owned());
+        headers.insert(
+            CONTENT_TYPE.as_str().to_string(),
+            "application/x-www-form-urlencoded".to_owned(),
+        );
+        #[derive(Deserialize, PartialEq, Debug)]
+        struct Payload {
+            foo: String,
+            baz: usize,
+        }
+        let gwr: GatewayRequest = GatewayRequest {
+            path: "/foo".into(),
+            http_method: "GET".into(),
+            headers: headers,
+            body: Some("foo=bar&baz=2".into()),
+            ..Default::default()
+        };
+        let actual = HttpRequest::from(gwr);
+        let payload: Option<Payload> = actual.payload().unwrap_or_else(|_| None);
+        assert_eq!(
+            payload,
+            Some(Payload {
+                foo: "bar".into(),
+                baz: 2
+            })
+        )
+    }
+
+    #[test]
+    fn requests_have_form_post_parseable_payloads_for_hashmaps() {
+        let mut headers = HashMap::new();
+        headers.insert(HOST.as_str().to_string(), "www.rust-lang.org".to_owned());
+        headers.insert(
+            CONTENT_TYPE.as_str().to_string(),
+            "application/x-www-form-urlencoded".to_owned(),
+        );
+        let gwr: GatewayRequest = GatewayRequest {
+            path: "/foo".into(),
+            http_method: "GET".into(),
+            headers: headers,
+            body: Some("foo=bar&baz=2".into()),
+            ..Default::default()
+        };
+        let actual = HttpRequest::from(gwr);
+        let mut expected = HashMap::new();
+        expected.insert("foo".to_string(), "bar".to_string());
+        expected.insert("baz".to_string(), "2".to_string());
+        let payload: Option<HashMap<String, String>> = actual.payload().unwrap_or_else(|_| None);
+        assert_eq!(payload, Some(expected))
+    }
+
+    #[test]
+    fn requests_have_json_parseable_payloads() {
+        let mut headers = HashMap::new();
+        headers.insert(HOST.as_str().to_string(), "www.rust-lang.org".to_owned());
+        headers.insert(
+            CONTENT_TYPE.as_str().to_string(),
+            "application/json".to_owned(),
+        );
+        #[derive(Deserialize, PartialEq, Debug)]
+        struct Payload {
+            foo: String,
+            baz: usize,
+        }
+        let gwr: GatewayRequest = GatewayRequest {
+            path: "/foo".into(),
+            http_method: "GET".into(),
+            headers: headers,
+            body: Some(r#"{"foo":"bar", "baz": 2}"#.into()),
+            ..Default::default()
+        };
+        let actual = HttpRequest::from(gwr);
+        let payload: Option<Payload> = actual.payload().unwrap_or_else(|_| None);
+        assert_eq!(
+            payload,
+            Some(Payload {
+                foo: "bar".into(),
+                baz: 2
+            })
+        )
     }
 }